Split up compositional multiphase solver
