+++ src/coreComponents/constitutive/fluid/multifluid/MultiFluidSelector.hpp
@@ -43,16 +43,15 @@ void constitutiveUpdatePassThru( MultiFluidBase const & fluid,
                                BlackOilFluid,
                                CompositionalTwoPhasePengRobinsonConstantViscosity,
                                CompositionalTwoPhaseSoaveRedlichKwongConstantViscosity,
-//                               CompositionalTwoPhasePengRobinsonLBCViscosity,
-//                               CompositionalTwoPhaseSoaveRedlichKwongLBCViscosity,
+                               CompositionalTwoPhasePengRobinsonLBCViscosity,
+                               CompositionalTwoPhaseSoaveRedlichKwongLBCViscosity,
 #ifdef GEOSX_USE_PVTPackage
                                CompositionalMultiphaseFluidPVTPackage,
 #endif
                                CO2BrinePhillipsFluid,
                                CO2BrineEzrokhiFluid,
-                               CO2BrinePhillipsThermalFluid
-//                               ,CO2BrineEzrokhiThermalFluid   "Uncommenting this will lead to compiler segfault. Need to split compilation
-// units for all the options"
+                               CO2BrinePhillipsThermalFluid,
+                               CO2BrineEzrokhiThermalFluid
                                >::execute( fluid, std::forward< LAMBDA >( lambda ) );
 }
 
@@ -64,16 +63,15 @@ void constitutiveUpdatePassThru( MultiFluidBase & fluid,
                                BlackOilFluid,
                                CompositionalTwoPhasePengRobinsonConstantViscosity,
                                CompositionalTwoPhaseSoaveRedlichKwongConstantViscosity,
-//                               CompositionalTwoPhasePengRobinsonLBCViscosity,
-//                               CompositionalTwoPhaseSoaveRedlichKwongLBCViscosity,
+                               CompositionalTwoPhasePengRobinsonLBCViscosity,
+                               CompositionalTwoPhaseSoaveRedlichKwongLBCViscosity,
 #ifdef GEOSX_USE_PVTPackage
                                CompositionalMultiphaseFluidPVTPackage,
 #endif
                                CO2BrinePhillipsFluid,
                                CO2BrineEzrokhiFluid,
-                               CO2BrinePhillipsThermalFluid
-                               //,CO2BrineEzrokhiThermalFluid   "Uncommenting this will lead to compiler segfault. Need to split compilation
-// units for all the options"
+                               CO2BrinePhillipsThermalFluid,
+                               CO2BrineEzrokhiThermalFluid
                                >::execute( fluid, std::forward< LAMBDA >( lambda ) );
 }
