+++ src/coreComponents/unitTests/constitutiveTests/CMakeLists.txt
@@ -23,6 +23,7 @@ set( gtest_pvt_xmls
      testPVT.xml
      testPVT_CO2Brine.xml
      testPVT_CO2BrineTables.xml
+     testPVT_Compositional.xml
      testPVT_PhaseComposition.xml
    )
